Dimension mismatch in analyzeshorttraj.m
In the updated FSMIA, the trajectories are in the column form. This means that when the data are concatenated together in this file to traj_complete, the dimensions don't match. I suggest transposing the initial assignment of the molecular indices to trajStat.traj at line 25.

I also don't need the printed numbers for debugging, so I commented them out. But if you still need them, nevermind!